back_up

+++ flask_server/community/community.py
@@ -19,27 +19,39 @@ class check_selfintro_ab(Resource):
     def get(self):
         return "this is community main page\n"
 
-#/sef_intro/view_comment/?page=<int>&per_page=<int>&question_uuid=<string>
 @api.route('/view_comment/')
 class view_comment(Resource):
     def get(self):
         page = int(request.args.get('page'))
         per_page = int(request.args.get('per_page'))
         question_uuid = request.args.get('ques_uuid')
+        viewer_uuid = request.args.get('user_uuid')
         page_offset = (page-1)*per_page
         
         comment_records = db.session.query(CommunityComment).\
             filter(CommunityComment.common_ques == question_uuid).\
                 order_by(CommunityComment.recommendation.desc()).\
                     offset(page_offset).limit(int(per_page)).all()
+        
         ret = []
         if(not comment_records):
             return ret
         
         for cr in comment_records:
+            ur = User.query.get(cr.user_uuid)
+            cr_uuid = CommentRecommendation.query.filter(CommentRecommendation.cc_uuid == cr.cc_uuid, \
+                CommentRecommendation.user_uuid == viewer_uuid).first()
+            
+            if(cr_uuid):
+                vw_liked = True
+            else:
+                vw_liked = False
+                
             ret.append({"cc_uuid":cr.cc_uuid,\
                 "user_uuid":cr.user_uuid, "common_ques":cr.common_ques, \
-                        "comment":cr.comment, "date":int(cr.date.timestamp()*1000), "recommendation":cr.recommendation})
+                        "comment":cr.comment, "date":int(cr.date.timestamp()*1000), \
+                            "recommendation":cr.recommendation, "name": ur.name, \
+                                "email":ur.email, "viewer_liked": vw_liked})
         
         #cc_uuid, user_uuid, common_ques, comment, date, recommendation
         return jsonify(ret)
@@ -66,8 +78,11 @@ class recommendation(Resource):
                 CommentRecommendation.user_uuid == user_uuid).first()
         check_user_record = db.session.query(User).get(user_uuid)
         
+        if(not cc_record):
+            return -2
+        
         if(not check_user_record):
-            return -2 # not user
+            return -1 # not user
         elif(check_cr_record):
             cc_record.recommendation -= 1
             db.session.delete(check_cr_record)
@@ -79,43 +94,54 @@ class recommendation(Resource):
         db.session.commit()
         return cc_record.recommendation
     
-
 @api.route('/comment/')
 class comment(Resource):
     def post(self):
         try:
-            cc_uuid = request.args.get('cc_uuid')
+            cc_uuid = uuid.uuid1()
             ques_uuid = request.get_json()['ques_uuid']
             user_uuid = request.get_json()['user_uuid']
             comment = request.get_json()['comment']
             cc_record = db.session.query(CommunityComment).get(cc_uuid)
             if(cc_record):
-                return 0
-            new_cc_record = CommunityComment(cc_uuid, user_uuid, ques_uuid, comment)
-            db.session.add(new_cc_record)
+                cc_record.comment = comment
+            else:
+                new_cc_record = CommunityComment(cc_uuid, user_uuid, ques_uuid, comment)
+                db.session.add(new_cc_record)
             db.session.commit()
         except:
-            return 0
-        return 1
+            return None
+        user_record = User.query.get(user_uuid)
+        return jsonify({"user_uuid": user_uuid, "cc_uuid": cc_uuid, "common_ques": ques_uuid,\
+                "comment": comment, "date": int(datetime.now(pytz.timezone('Asia/Seoul')).timestamp() * 1000), \
+                    "recommendation": 0, "name": user_record.name, "email": user_record.email, \
+                        "viewer_liked": False})
     def put(self):
         try:
             cc_uuid = request.args.get('cc_uuid')
             comment = request.get_json()['comment']
             cc_record = db.session.query(CommunityComment).get(cc_uuid)
             cc_record.comment = comment
+            cc_record.date = datetime.now(pytz.timezone('Asia/Seoul'))
             db.session.commit()
         except:
-            return 0
+            return None
+        
+        user_record = User.query.get(cc_record.user_uuid)
+        return jsonify({"user_uuid": cc_record.user_uuid, "cc_uuid": cc_uuid, "common_ques": cc_record.common_ques,\
+                "comment": comment, "date": int(datetime.now(pytz.timezone('Asia/Seoul')).timestamp() * 1000), \
+                    "recommendation": 0, "name": user_record.name, "email": user_record.email, \
+                        "viewer_liked": False})
     def delete(self):
         cc_uuid = request.args.get('cc_uuid')
+        user_uuid = request.args.get('user_uuid')
         cc_record = db.session.query(CommunityComment).get(cc_uuid)
+        if(not cc_record or cc_record.user_uuid != user_uuid):
+            return 0
         cr_records = db.session.query(CommentRecommendation).filter(CommentRecommendation.cc_uuid == cc_uuid).all()
         for cr in cr_records:
             db.session.delete(cr)
         db.session.commit()
-        print(1)
-        if(not cc_record):
-            return 0
         db.session.delete(cc_record)
         db.session.commit()
         return 1
@@ -124,30 +150,44 @@ class comment(Resource):
 class my_comment(Resource):
     def get(self):
         user_uuid = request.args.get('user_uuid')
-        records = CommunityComment.query.filter(CommunityComment.user_uuid == user_uuid)
+        records = CommunityComment.query.filter(CommunityComment.user_uuid == user_uuid).all()
+        user = User.query.get(user_uuid)
         
         ret = []
-        for record in records:
-            question = CommonQue.query.get(record.common_ques).question
-            ret.append({"cc_uuid": record.cc_uuid, "ques": question,\
-                "comment": record.comment, "date": int(record.date.timestamp()) * 1000, \
-                    "recommendation": record.recommendation})
         
-        return jsonify(ret)
+        if(not records):
+            return None
+        
+        for record in records:
+            cr = CommentRecommendation.query.filter(CommentRecommendation.cc_uuid == record.cc_uuid, \
+                CommentRecommendation.user_uuid == user_uuid).first()
+            if(cr): vw_liked = True
+            else: vw_liked = False
+            ret.append({"user_uuid": user_uuid, "cc_uuid": record.cc_uuid, "common_ques": record.common_ques,\
+                "comment": record.comment, "date": int(record.date.timestamp()) * 1000, \
+                    "recommendation": record.recommendation, "name": user.name, "email": user.email, \
+                        "viewer_liked": vw_liked})
+        
+        return jsonify(ret[0])
 
 @api.route('/view_my_ques_comment/')
 class my_comment(Resource):
     def get(self):
         user_uuid = request.args.get('user_uuid')
         ques_uuid = request.args.get('ques_uuid')
-        records = CommunityComment.query.filter(CommunityComment.user_uuid == user_uuid, \
-            CommunityComment.common_ques)
+        record = CommunityComment.query.filter(CommunityComment.user_uuid == user_uuid, \
+            CommunityComment.common_ques).first()
+        user = User.query.get(user_uuid)
         
-        ret = []
-        for record in records:
-            question = CommonQue.query.get(record.common_ques).question
-            ret.append({"cc_uuid": record.cc_uuid, "ques": question,\
-                "comment": record.comment, "date": int(record.date.timestamp()) * 1000, \
-                    "recommendation": record.recommendation})
+        cr = CommentRecommendation.query.filter(CommentRecommendation.cc_uuid == ques_uuid, \
+                CommentRecommendation.user_uuid == user_uuid).first()
+            
+        if(cr):
+            vw_liked = True
+        else:
+            vw_liked = False
         
-        return jsonify(ret)
+        return jsonify({"user_uuid": user_uuid, "cc_uuid": record.cc_uuid, "common_ques": record.common_ques,\
+                "comment": record.comment, "date": int(record.date.timestamp() * 1000), \
+                    "recommendation": record.recommendation, "name": user.name, "email": user.email, \
+                        "viewer_liked": vw_liked})
